Upgrade to Next 13.3, fix indicator not updating
This commit upgrades to the latest version of Next.js, including
migrating from @next/font to next/font, using the new `metadata` fields,
and removing unused dependencies.

Along the way, I noticed that the "Now" indicator wasn't updating
because we were now statically rendering the site once on the server-
side. This commit fixes that by using the `revalidate` field to force
the page to update once every 21 hours.